Pull networking CVE fixes from Steffen Jaeckel, KVM bug fix from Juergen
Gross, CVE metadata update from Michal Hocko and SCSI fixes from Lee
Duncan

suse-commit: 92663b2b1774f6270538b1b0969fa100fdbe7a0e

+++ arch/x86/kvm/mmu/mmu.c
@@ -3878,7 +3878,7 @@ static bool try_async_pf(struct kvm_vcpu *vcpu, bool prefault, gfn_t gfn,
  * root was invalidated by a memslot update or a relevant mmu_notifier fired.
  */
 static bool is_page_fault_stale(struct kvm_vcpu *vcpu, kvm_pfn_t pfn,
-				hva_t hva, int mmu_seq)
+				hva_t hva, unsigned long mmu_seq)
 {
 	struct kvm_mmu_page *sp = to_shadow_page(vcpu->arch.mmu->root_hpa);
